Add missing addition in VarIntTypeInt.ftl and unfold stream processing for better debugging

+++ backend/src/main/java/com/bakdata/conquery/models/types/parser/specific/VarIntParser.java
@@ -51,6 +51,8 @@ public class VarIntParser extends Parser<Integer> {
 			);
 		}
 
-		return new Decision<>(value -> value - minValue, new VarIntTypeInt(minValue, maxValue));
+
+
+		return new Decision<>(value -> (int) (value - minValue), new VarIntTypeInt(minValue, maxValue));
 	}
 }
